+++ plugin/src/mosaic/core/particleLinking/ParticleLinker_MarkKittisopikul.java
@@ -57,7 +57,7 @@ public class ParticleLinker_MarkKittisopikul{
 			for(n = 0; n < curr_linkrange; n++) {
 				IJ.showStatus("Linking Frame " + (m+1) + ": initialized " + nop + "Particles");
 				IJ.showProgress(n,curr_linkrange);
-				max_cost = (double)(n + 1) * displacement * (double)(n + 1) * displacement;
+				max_cost = (n + 1) * displacement * (n + 1) * displacement;
 
 				nop_next = frames[m + (n + 1)].getParticles().size();
 //				System.out.println("nop: " + nop);
