+++ plugin/src/mosaic/paramopt/cma/PrintfFormat.java
@@ -2152,7 +2152,7 @@ public class PrintfFormat {
         }
       }
       else
-        sx = Integer.toString((int)x,16);
+        sx = Integer.toString(x,16);
       return printXFormat(sx);
     }
     /**
@@ -2447,7 +2447,7 @@ public class PrintfFormat {
         }
       }
       else
-        sx = Integer.toString((int)x,8);
+        sx = Integer.toString(x,8);
       return printOFormat(sx);
     }
     /**
